pkg/prometheus: Move config to independent package
Avoid packages that only need the config having to pull all dependencies.
For instance, this avoid kubectl-gadget importing
"github.com/inspektor-gadget/inspektor-gadget/pkg/runtime/local".
